fixed layout and added loading
